Allow to pass the values for platform and platform-params via env vars.
On some cases it is useful to be able to specify the Cog platform plugin
and the platform parameters via environment variables.

For example: bots running CI tests may be executing the same command to
start Cog but each bot can be configured to use a different platform
plugin via enviroment variables.

Another example: The script `run-benchmark` from WebKit doesn't allow
the user to pass custom command-line parameters to the browser, so
with this environment variables the user should be able to select
the platform parameters for Cog via the environment.